requirements.txt
python-telegram-bot[job-queue]>=21.0,<22
python-dotenv>=1.0.0
